Modify output table in randomly stock selection experiment
Feat: Modify the format of the result table. 1.The cell color would be red if the stock goes up, green if the stock goes down. 2.The directional triangle indicates whether stock goes up or down. 3.Change the order of columns, now that date is the first column. 4.Change the column name to Chinese