perf: optimize performance when checking the valid date in random stock selection experiment
This commit introduce an alternative method to check the valid date in random stockselection experiment. The original method would pull the whole data frame and check the first and the last row to determine whether this stock is valid. New method only pull the data frame that has recorded the valid date for each stock.